+++ libphx/src/MemStack.rs
@@ -15,7 +15,7 @@ pub struct MemStack {
 #[no_mangle]
 pub unsafe extern "C" fn MemStack_Create(mut capacity: u32) -> *mut MemStack {
     let mut this: *mut MemStack = MemAlloc(::core::mem::size_of::<MemStack>()) as *mut MemStack;
-    (*this).size = 0_i32 as u32;
+    (*this).size = 0_u32;
     (*this).capacity = capacity;
     (*this).data = MemAlloc(capacity as usize);
     this
@@ -46,7 +46,7 @@ pub unsafe extern "C" fn MemStack_Alloc(
 
 #[no_mangle]
 pub unsafe extern "C" fn MemStack_Clear(mut this: *mut MemStack) {
-    (*this).size = 0_i32 as u32;
+    (*this).size = 0_u32;
 }
 
 #[no_mangle]
